Display fees in edit modals and fix isolated fee calc
