+++ state/futures/selectors.ts
@@ -1491,11 +1491,3 @@ export const selectMarketSuspended = createSelector(
 	selectMarketInfo,
 	(marketInfo) => marketInfo?.isSuspended
 );
-
-export const selectClosePositionOrderFee = createSelector(
-	(state: RootState) => state.futures.closePositionOrderFee,
-	wei
-);
-
-export const selectClosePositionOrderFeeError = (state: RootState) =>
-	state.futures.queryStatuses.closePositionOrderFee.error;
